moved rescaled mask
